+++ apps/backend/src/libs/hooks/hooks.ts
@@ -1,3 +1,4 @@
 export { checkByParameterIfNotTheSameUser } from "./check-by-parameter-if-not-the-same-user/check-by-parameter-if-not-the-same-user.hook.js";
 export { checkIdEquality } from "./check-id-equality/check-id-equality.hook.js";
 export { checkUserPermissions } from "./check-user-permissions/check-user-permissions.hook.js";
+export { checkUserSubscription } from "./check-user-subscription/check-user-subscription.hook.js";