+++ apps/frontend/src/libs/components/table/libs/helpers/helpers.ts
@@ -1,0 +1,1 @@
+export { getCellProperties } from "./get-cell-properties.helper.js";