+++ apps/frontend/src/modules/user-courses/slices/user-courses.slice.ts
@@ -2,7 +2,6 @@ import { createSlice } from "@reduxjs/toolkit";
 
 import { DataStatus } from "~/libs/enums/enums.js";
 import { type ValueOf } from "~/libs/types/types.js";
-import { type CourseDto } from "~/modules/courses/courses.js";
 
 import { type UserCourseResponseDto } from "../libs/types/types.js";
 import {
@@ -14,7 +13,7 @@ import {
 
 type State = {
 	addCourseDataStatus: ValueOf<typeof DataStatus>;
-	commonCourses: CourseDto[];
+	commonCoursesIds: number[];
 	commonDataStatus: ValueOf<typeof DataStatus>;
 	dataStatus: ValueOf<typeof DataStatus>;
 	myCourses: UserCourseResponseDto[];
@@ -25,7 +24,7 @@ type State = {
 
 const initialState: State = {
 	addCourseDataStatus: DataStatus.IDLE,
-	commonCourses: [],
+	commonCoursesIds: [],
 	commonDataStatus: DataStatus.IDLE,
 	dataStatus: DataStatus.IDLE,
 	myCourses: [],
@@ -69,7 +68,7 @@ const { actions, name, reducer } = createSlice({
 			state.dataStatus = DataStatus.REJECTED;
 		});
 		builder.addCase(loadCommonCourses.fulfilled, (state, action) => {
-			state.commonCourses = action.payload;
+			state.commonCoursesIds = action.payload;
 			state.commonDataStatus = DataStatus.FULFILLED;
 		});
 		builder.addCase(loadCommonCourses.pending, (state) => {