+++ apps/frontend/src/pages/management/libs/components/courses-tab/libs/components/courses-table/libs/types/types.ts
@@ -1,0 +1,1 @@
+export { type CoursesTableRow } from "./courses-table-row.type.js";