+++ packages/shared/src/modules/user-courses/libs/types/types.ts
@@ -1,1 +1,2 @@
+export { type CommonCoursesResponseDto } from "./common-courses-response-dto.type.js";
 export { type UserCourseResponseDto } from "./user-course-response-dto.type.js";